Compared to the mainline, this code runs at about the same speed in CPython 3.11.3 but is significantly faster in PyPy 7.3.12.

```
cpython_old: 24.7744 seconds
cpython_new: 24.2009 seconds
2.37% faster

pypy_old: 4.13073 seconds
pypy_new: 1.56451 seconds
164% faster
```

Benchmark code:
```python
import sympy
u = list(sympy.sieve.totientrange(10**7, 2 * 10**7))
``` 

<!-- BEGIN RELEASE NOTES -->
* ntheory
  * Speed up `totientrange`.
<!-- END RELEASE NOTES -->