@yuanyelele 2023-09-05
Also, I think we should raise an exception (maybe `IndexError`) when `totientrange(a, b)` is called with `a <= 0 < b`.